Update weaver resolver to use cache that remembers resolved schema by schema-url.
