+++ crates/weaver_semconv/Cargo.toml
@@ -31,6 +31,7 @@ globset.workspace = true
 itertools.workspace = true
 log.workspace = true
 url.workspace = true
+semver.workspace = true
 
 glob.workspace = true
 jsonschema = "0.46.4"        # JSON Schema validation used to enhance error messages
